283 Return to old exit priority schema

+++ plasma_framework/contracts/mocks/exits/SpyPlasmaFrameworkForExitGame.sol
@@ -24,7 +24,7 @@ contract SpyPlasmaFrameworkForExitGame is PlasmaFramework {
     }
 
     /** override for test */
-    function enqueue(address _token, uint64 _exitableAt, TxPosLib.TxPos calldata _txPos, uint192 _exitId, IExitProcessor _exitProcessor)
+    function enqueue(address _token, uint64 _exitableAt, TxPosLib.TxPos calldata _txPos, uint160 _exitId, IExitProcessor _exitProcessor)
         external
         returns (uint256)
     {
